Refactor the mapping of common in contact controller to be more obvious and refactor company formatting to fix a weird mocha dependency bug
